Refactor file handling
